scraper/linkedin_scraper.py
import re
import urllib.parse
from typing import List, Dict, Optional
from bs4 import BeautifulSoup
from datetime import datetime
import os
from .base_scraper import BaseScraper

class LinkedInScraper(BaseScraper):
    """Scraper for LinkedIn posts and ads"""
    
    def __init__(self):
        super().__init__('linkedin')
        self.base_url = 'https://www.linkedin.com'
        self.search_url = 'https://www.linkedin.com/search/results/content/'
        
        # LinkedIn specific headers
        self.session.headers.update({
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
            'Accept-Encoding': 'gzip, deflate, br',
            'DNT': '1',
            'Connection': 'keep-alive',
            'Upgrade-Insecure-Requests': '1'
        })
    
    def build_search_url(self, query: str, filters: Dict = None) -> str:
        """Build LinkedIn search URL with query and filters"""
        params = {
            'keywords': query,
            'origin': 'GLOBAL_SEARCH_HEADER',
            'sid': 'b7J'
        }
        
        if filters:
            # Add time filter (past week, month, etc.)
            if 'datePosted' in filters:
                params['datePosted'] = filters['datePosted']
            
            # Add content type filter
            if 'resultType' in filters:
                params['resultType'] = filters['resultType']
        
        return f"{self.search_url}?{urllib.parse.urlencode(params)}"
    
    def scrape(self, search_query: Optional[str] = None) -> List[Dict]:
        """Scrape LinkedIn for posts mentioning our keywords"""
        results = []
        
        # If no specific query, search for each keyword
        if not search_query:
            all_keywords = []
            for category_keywords in self.keywords.values():
                all_keywords.extend(category_keywords)
            search_query = ' OR '.join([f'"{kw}"' for kw in all_keywords])
        
        print(f"Searching LinkedIn for: {search_query}")
        
        # Note: LinkedIn requires authentication for most content
        # This is a simplified version that works with public content
        # For full functionality, you'd need to implement login
        
        # Try Google search as a workaround for public LinkedIn content
        results.extend(self._scrape_via_google(search_query))
        
        # Save results to database
        if results:
            self.save_results(results)
        
        return results
    
    def _scrape_via_google(self, query: str) -> List[Dict]:
        """Scrape LinkedIn content via Google search"""
        results = []
        
        # Build Google search query
        google_query = f'site:linkedin.com/posts OR site:linkedin.com/pulse "{query}"'
        google_url = f"https://www.google.com/search?q={urllib.parse.quote(google_query)}&num=50"
        
        # Get search results
        content = self.get_page_content(google_url)
        if not content:
            print("Failed to get Google search results")
            return results
        
        soup = BeautifulSoup(content, 'html.parser')
        
        # Parse Google search results
        for result in soup.find_all('div', class_='g'):
            link_elem = result.find('a')
            if not link_elem:
                continue
            
            url = link_elem.get('href', '')
            if 'linkedin.com' not in url:
                continue
            
            # Extract snippet
            snippet_elem = result.find('span', class_='aCOpRe')
            snippet = snippet_elem.text if snippet_elem else ''
            
            # Check if snippet contains our keywords
            keywords_found = self._extract_keywords(snippet)
            if not keywords_found:
                continue
            
            # Extract title
            title_elem = result.find('h3')
            title = title_elem.text if title_elem else 'LinkedIn Post'
            
            # Try to extract author from title or URL
            author = self._extract_author_from_title(title)
            
            # Determine post type
            post_type = 'article' if '/pulse/' in url else 'post'
            
            results.append({
                'platform': self.platform,
                'post_type': post_type,
                'url': url,
                'author': author,
                'author_title': None,  # Would need to visit the page
                'content': snippet,
                'timestamp': None,  # Would need to visit the page
                'keywords_found': keywords_found,
                'likes': None,
                'comments': None,
                'shares': None,
                'company': None,
                'location': None,
                'is_verified': False
            })
        
        print(f"Found {len(results)} LinkedIn results via Google")
        return results
    
    def _extract_author_from_title(self, title: str) -> Optional[str]:
        """Try to extract author name from post title"""
        # Common patterns in LinkedIn titles
        patterns = [
            r'^(.+?) on LinkedIn:',
            r'^(.+?) posted on LinkedIn',
            r'^(.+?) shared',
            r'^Post by (.+?)$'
        ]
        
        for pattern in patterns:
            match = re.search(pattern, title)
            if match:
                return match.group(1).strip()
        
        return None
    
    def parse_post(self, post_element) -> Optional[Dict]:
        """Parse individual LinkedIn post element"""
        # This would be implemented if we had direct access to LinkedIn HTML
        # For now, this is handled in _scrape_via_google
        pass
    
    def scrape_linkedin_ads(self) -> List[Dict]:
        """Scrape LinkedIn ads (requires different approach)"""
        # LinkedIn ads are typically not publicly accessible
        # Would require LinkedIn Ads API or authenticated scraping
        print("LinkedIn ads scraping requires authentication or API access")
        return []