requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
pandas==2.2.0
flask==3.0.2
flask-cors==4.0.0
sqlalchemy==2.0.27
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.1.0
html5lib==1.1
webdriver-manager==4.0.1
playwright==1.41.2
fake-useragent==1.5.1
cloudscraper==1.2.71